Separate test_json
